Add complete solution for Yakuso
The solver now uses two separate lists: a board with 0=no value and 1=max value and a separate list for keeping the max value per row. This proved to be the most simple solution.

We also add constraints for the sums so that we could *in theory* generate valid Yakusos ourselves.